new runtime: rename thread.h thread_info.h
